feat(memory): improve virtual memory paging subsystem
- Added 2MB and 1GB huge page translation support in `src/memory/paging.rs`.
- Implemented virtual memory areas (VMAs) for simulated on-demand paging.
- Implemented Clock (second-chance) page replacement tracking algorithm to prevent loop spins.
- Added comprehensive unit tests for all virtual memory, huge page, and clock replacement logic.
- Synchronized all strategic wiki documents and timelines.

+++ src/boot/pci.rs
@@ -6,7 +6,7 @@
 extern crate alloc;
 use alloc::vec::Vec;
 
-pub const PCI_MAX_BUS: u8 = 256;
+pub const PCI_MAX_BUS: u16 = 256;
 pub const PCI_MAX_DEVICE: u8 = 32;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
